Move value functions to values module
